+++ src/main/java/com/example/gestuon_promotions_marjan/DAO/IDAO.java
@@ -9,7 +9,7 @@ public interface IDAO<T> {
 
     public List<T> findByDesignation(String mc);
 
-    public T findByid(long id);
+    public T findByid(int id);
 
     public T update(T T);
